Small fixes and cleanups
